move fizzbuzztree to bst
