+++ data_structures/binary_search_tree/README.md
@@ -13,6 +13,14 @@ Create a Class for a BST, which is aware of the root of the tree as root
 Ensure that your class and any subsequent methods are properly tested, and that your test coverage is above 80%.
 
 
+### Whiteboard Challenge - Conduct “FizzBuzz” on a tree while traversing through it. Change the values of each of the nodes dependent on the current node’s value
+
+- Write a function called FizzBuzzTree which takes a tree as an argument.
+- Without utilizing any of the built-in methods available to your language, determine weather or not the value of each node is divisible by 3, 5 or both, and change the value of each of the nodes respectively. Return the tree with it’s news values.
+
+![whiteboard](../../assets/fizzbuzztree.jpg)
+
+
 ### Whiteboard Challenge - Breadth-First Traversal
 
 Write a function called breadthFirstTraversal which takes a Binary Tree as its unique input. Without utilizing any of the built-in methods available to your language, traverse the input tree using a Breadth-first approach; print every visited node’s value.
