+++ data_structures/binary_search_tree/test_fizzbuzztree.py
@@ -1,8 +1,9 @@
-from fizzbuzztree import BST, fizzbuzztree
+from .bst import BST
+from .fizzbuzztree import fizzbuzztree
 
 
 def test_fizzbuzztree_balanced():
-    """test fizz buzz on balanced tree"""
+    """Test fizz buzz on balanced tree."""
     balanced_bst = BST([10, 7, 3, 16, 15, 8, 20])
     fizzbuzztree(balanced_bst)
     assert balanced_bst.root.val == 'buzz'
@@ -11,7 +12,7 @@ def test_fizzbuzztree_balanced():
 
 
 def test_fizzbuzztree_right():
-    """test fizz buzz on right-heavy tree"""
+    """Test fizz buzz on right-heavy tree."""
     right_bst = BST([1, 3, 5, 7, 9, 15, 20, 21, 26])
     fizzbuzztree(right_bst)
     assert right_bst.root.right.val == 'fizz'
@@ -21,7 +22,7 @@ def test_fizzbuzztree_right():
 
 
 def test_fizzbuzztree_left():
-    """test fizz buzz on left-heavy tree"""
+    """Test fizz buzz on left-heavy tree."""
     left_bst = BST([30, 10, 8, 6, 3])
     fizzbuzztree(left_bst)
     assert left_bst.root.val == 'fizz buzz'
